Bug 241414: JPA 2.0 Add support for a Uni-directional OneToMany mapping without Join Table The patch adds target optimistic locking support
The target version is incremented in case the target is added to (or removed
from) the source; also in case the source is deleted. That means that the
target version is always incremented if foreign key is changed. Note that on
source deletion the target indirection is triggered, and its possible to get
OptimisticLocking exception because the target has been changed by another thread.


Former-commit-id: aa2c94031f19b2a5c705a6d3269c101b5f9441fe