Bug 249037: JPA 2.0 persisting list item index Correction.
Target foreign key and list order column must be in the same table, so either
specify list order column with target foreign key table or don't specify list
order column's table at all - it will be defaulted to foreign key table.
The correction fixes the case when the table is not specified.


Former-commit-id: 15f058d6b247c66c643109f7d6a6bce8c9fcb5aa