Bug 305713 - OuterJoinExpressionHolder linearization order can be incorrect during ansi joins
Reopening the bug.
The wrong ordering of ansi outer joins still happens when the secondary table is
the source of auto join.

The new patch overrides the core changes (the only altered core class is
SQLSelectStatement). The previously suggested test for the issue is kept (in
the previous patch). The new test is added that reproduses the issue:
        ExpressionBuilder eb = query.getExpressionBuilder();        
        Expression expression = eb.getAllowingNull("address");
        query.addJoinedAttribute(expression);

        expression = eb.get("manager");
        query.addJoinedAttribute(expression);
        expression = expression.getAllowingNull("address");
        query.addJoinedAttribute(expression);

        expression = eb.getAllowingNull("hugeProject");
        query.addJoinedAttribute(expression);
        expression = expression.getAllowingNull("evangelist");
        query.addJoinedAttribute(expression);

where a new subclass HugeProject of LargeProject was added to Employee test
model. 
Employee defines a new "hugeProject" OneToOne relation to HugeProject.
HugeProject defines a new "evangelist" OneToOne relation to Employee.
The two relations are independent.


Former-commit-id: 2c11dd1acdb5ef58f09c4cbdfbcd6837857793e2