Bug 244124 - Add Nested FetchGroup
Main patch - take 3.
There are some things not done yet:
1) merging in cache still uses "one-storey" union;
2) fetch joins

All the fetch groups assigned to objects now are of type EntityFetchGroup.
It's a flat (not nested) fetch group that only reflects the state of its owner
object.
A single EntityFetchGroup may be shared by many objects, and it is immutable.
An EntityFetchGroup usually not created directly but rather looked up on the
FetchGroupmanager (keyed by set of its attributes).

When a value is set into not fetched attribute in serialized object (the one
that doesn't have a session) then a new EntityFetchGroup created to reflect the
new object's fetch state.

Deprecated partial attributes
Added LOAD_GROUP_ATTRIBUTE QueryHint.
Added "load" method to JpaEntityManager.


Former-commit-id: bcda2368e0de14b34050e1bc552179c293629cec