Bug 244124 - Add Nested FetchGroup Correction patch after James' review.
James' comments:

- addPartialAttribute, did you also deprecate the one that takes an Expression?

- ObjectLevelReadQuery, addSelectionFieldsForJoinedExpression, code is odd,
cast to Vector<DatabaseField> but then casts to Object and allows Expression?

- getFetchGroupSelectionFields, in general try to avoid using Vector (List).

- shouldReadMapping, short circuit was removed?



Some more comments from refreshing my view.

- FetchGroupManager, has a bunch of public methods that should be marked
internal

- Should have a test for embeddable attributes, we should probably throw some
kind of error if we don't support this, also aggregate collections.

- ForeignReferenceMapping, batchValueFromRow, the setting of the fetch group
should be in the prepare of the nested batch query, not cloning it and
repreparing it.

- FetchGroup, set/shouldLoad should have a user comment explaining what is does

- LoadGroup, should avoid using DescriptorIterator as performance issue and not
polymorphic, should dispatch to ObjectBuilder/mappings

- ObjectLevelReadQuery, setting of fetch group should be in mapping, not
cloning and repreparing query in query

- ClassWeaver, does not look like fetch groups are weaved if change tracking is
not weaved?

- MethodWeaver, same thing

- QuerySQLTracker uses tabs, also the class already exists in the testing
framework package, so no need to define another


The patch corrects most of the problems mentioned in these comments.


Former-commit-id: 52ada1f181b9cfc43ee357e7f90866785ef40867