Fix for Bug 322183 - NullPolicy & element with just attributes producing null
Added a new event to DeferredContentHandler called processEmptyElementWithAttributes().  By default this just calls processEmptyElement.  This is overridden in CompositeObjectMappingContentHandler where it is treated as processComplexElement().  

The processEmptyElementWithAttributes event is called when there are no sub-elements or characters and atleast one attribute that is not a namespace declaration or in the xsi namespace.

Code reviewed by:  Matt MacIvor



Former-commit-id: 9d64f6fa71a88b0b965ae472eeaaecccf21da2e3