Bug 257756 - problem with embedded mapping referencing an Entity overriden as an Embeddable
Reviewed by: Tom Ware

Tests: No, model change made to existing XML advanced model EmploymentPeriod (marked as Embeddable in XML). Adding an
@Entity to it caused a metadata processing validation exception since the startDate and endDate variables are not
decorated with the JPA required @Temporal. Meaning, the Employment class was being processed as an entity as well (and
apparent in the stack track). With this change, that exception goes away and all tests (FullRegressionTestSuite and
extended JPA) pass successfully.



Former-commit-id: bcb244847fc6dfa94f7b2d8e29ba61853be3681c