Fix for Bug 337874 - Memory Improvement - Reduce Map creation of SDOProperty.propertyValues
Code reviewed by:  David McCann

The most common open content property for Property is "xmlElement".  This was causing an entire map to be created just for this one value.  As an optimization this value has been pulled into its own instance variable.  The creation of the Map is delayed until another open content property is added.


Former-commit-id: 0f3c97b07ddccd49258e1a90cba1c24139d4c6a8