Phase 1 for Bug 343615 - Enhancement: Add ability to refresh JAXBContext based on metadata changes
Code reviewed by:  Matt MacIvor

Core - Fix (Phase 1):
- Introduced a new inner class on XMLContext called XMLContextState
- XMLContextState is responsible for holding onto all of an XMLContext's state,
when the refresh is implemented this will allow a single object to be swapped.
- All of an XMLContexts operations are based on an instance of XMLContextState.
 If the XMLContextState were to be swapped, the operation would continue
against the same XMLContextState it started with.

MOXy - Fix (Phase 1):
- Introduced a new inner class on JAXBContext called JAXBContextState
- JAXBContextState is responsible for holding onto all of an JAXBContext's
state, when the refresh is implemented this will allow a single object to be
swapped.
- All of a JAXBContexts operations are based on an instance of
JAXBContextState.  If the JAXBContextState were to be swapped, the operation
would continue against the same JAXBContextState it started with.
- Introduced a new inner class on JAXBContext called JAXBContextInput
- JAXBContextInput now contains the logic for creating a JAXBContext (atleast
the state), that was previously the responsibility of JAXBContextFactory.


Former-commit-id: 1da656ca959cd2c32715bd6f86cf03024b1dd8e5