Test Cases for Bug 336249 - Class property order doesn't allow superclass properties in it
MOXy is working as expected:

Scenario #1 - Parent class is marked with @XmlTrasient
If the parent class is marked with @XmlTransient then the parent's properties
can be included in the @XmlType(propOrder={...}) setting.

Scenario #2 - Parent class is not marked with @XmlTransient
If the parent class is not marked with @XmlTransient then then the parent's
properties can not be in propOrder set on the child class.  When inheritance is
mapped the parents properties must appear before the childs properties.


Former-commit-id: 42774dd8a3b8bbdc789939a27ea3b37219f3cee4