Bug 350487 - JPA 2.1 Specification defined support for Stored Procedure Calls
Removing the JPA 2.1 classes from the EclipseLink jar. We now load ALL jpa classes through statics to remain consistent. Also opens the door for the spec to remove annotations in the future (not that I expect they will) but, hey we'll still work out of the box if they do. As long as the new JPA annotation static referencing strategy is followed. :-)


Former-commit-id: 0e3b6e64f8e2acfdfb6fddb6e53aa01b86451c78