feature/bug 368365 - add DDL support to alter existing tables to add missing fields design doc http://wiki.eclipse.org/EclipseLink/DesignDocs/368365
Change to first attempt to create tables and only then read in the metadata if that fails.  This will avoid some of the issues where the table might exist in a different schema than the one it should be created in - potentially a common issue until 367887 is fixed.  
 
checked into 2.4
Reviewed by David Minsky


Former-commit-id: cbf949ba4ffd17c80fa743b80f2b01c1379ec0a2