Fix for Bug 340673 - Unexpected JAXBElement returned when unmarshalling with inheritance / XmlSeeAlso
Code reviewed by:  Rick Barkhouse

---

Resolution

There is a difference in how MOXy and the RI handle this use case.  As MOXy is able to round trip the following document and the RI is not, I believe the MOXy behaviour to be the correct behaviour for this use case.

<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
<animal xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:type="cat">
    <name>Fluffy</name>
    <sleepTime>10</sleepTime>
    <enemy>
        <name>Buster</name>
        <barkLevel>LOUD</barkLevel>
    </enemy>
</animal>

Note both MOXy and the RI handle the following document the same:

<cat>
    <name>Fluffy</name>
    <sleepTime>10</sleepTime>
    <enemy>
        <name>Buster</name>
        <barkLevel>LOUD</barkLevel>
    </enemy>
</cat>


Former-commit-id: e583d4c122f6beebba172897a7c4cf093197eeee